🚚 move: common 폴더 생성 후 관련 파일 이동
