+++ src/components/common/CommonGrid.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from "react";
 
 const CommonGrid = ({ children }: { children: ReactNode }) => (
-  <div className="grid grid-cols-2 md:grid-cols-4 lg:grid-cols-6 gap-4 w-full">
+  <div className="grid grid-cols-2 md:grid-cols-4 lg:grid-cols-5 gap-4 w-full">
     {children}
   </div>
 );
